# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/receipts/printers": {
      "get": {
        "operationId": "ReceiptsController_getPrinters",
        "summary": "Получить список доступных принтеров",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список принтеров получен"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/orders/{orderId}/receipt": {
      "post": {
        "operationId": "ReceiptsController_createReceipt",
        "summary": "Создать чек для заказа",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Чек успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "Невозможно создать чек"
          },
          "404": {
            "description": "Заказ не найден"
          },
          "409": {
            "description": "Чек для заказа уже существует"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/orders/{orderId}/receipt/compact": {
      "post": {
        "operationId": "ReceiptsController_createCompactReceipt",
        "summary": "Создать компактный чек для заказа",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Компактный чек успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "Невозможно создать чек"
          },
          "404": {
            "description": "Заказ не найден"
          },
          "409": {
            "description": "Чек для заказа уже существует"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/orders/{orderId}/receipt/standard": {
      "post": {
        "operationId": "ReceiptsController_createStandardReceipt",
        "summary": "Создать стандартный чек для заказа",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Стандартный чек успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "Невозможно создать чек"
          },
          "404": {
            "description": "Заказ не найден"
          },
          "409": {
            "description": "Чек для заказа уже существует"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts": {
      "get": {
        "operationId": "ReceiptsController_findAll",
        "summary": "Получить список всех чеков",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit (max 100)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список чеков получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReceiptResponse"
                }
              }
            }
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}": {
      "get": {
        "operationId": "ReceiptsController_findOne",
        "summary": "Получить метаданные чека",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Метаданные чека получены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "404": {
            "description": "Чек не найден"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}/pdf": {
      "get": {
        "operationId": "ReceiptsController_getPdf",
        "summary": "Скачать PDF чек",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF файл получен"
          },
          "404": {
            "description": "PDF файл не найден"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}/print": {
      "post": {
        "operationId": "ReceiptsController_printReceipt",
        "summary": "Отправить чек на печать",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "printer",
            "required": false,
            "in": "query",
            "description": "Имя принтера (по умолчанию системный)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Чек отправлен на печать"
          },
          "404": {
            "description": "Чек не найден"
          },
          "500": {
            "description": "Ошибка при печати"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}/void": {
      "post": {
        "operationId": "ReceiptsController_voidReceipt",
        "summary": "Void a receipt and unlock the order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt voided successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "Receipt is already voided"
          },
          "404": {
            "description": "Receipt not found"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/test-preview": {
      "post": {
        "operationId": "ReceiptsController_testPreview",
        "summary": "Generate a test receipt PDF preview using current user settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Test PDF generated"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}/regenerate": {
      "post": {
        "operationId": "ReceiptsController_regenerateReceiptPdf",
        "summary": "Принудительно регенерировать PDF чек",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF чек успешно регенерирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "404": {
            "description": "Чек не найден"
          },
          "500": {
            "description": "Ошибка при регенерации PDF"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}/share": {
      "post": {
        "operationId": "ReceiptsController_shareReceipt",
        "summary": "Generate a public share link for a receipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public link generated"
          },
          "404": {
            "description": "Receipt not found"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/receipts/{id}/share/revoke": {
      "post": {
        "operationId": "ReceiptsController_revokeShareReceipt",
        "summary": "Revoke public share link for a receipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public link revoked"
          },
          "404": {
            "description": "Receipt not found"
          }
        },
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/public/receipts/{token}": {
      "get": {
        "operationId": "PublicReceiptsController_getPublicReceipt",
        "summary": "Get public receipt data by token (no auth)",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt data returned"
          },
          "404": {
            "description": "Receipt not found or revoked"
          }
        },
        "tags": [
          "public-receipts"
        ]
      }
    },
    "/public/receipts/{token}/pdf": {
      "get": {
        "operationId": "PublicReceiptsController_getPublicReceiptPdf",
        "summary": "Download public receipt PDF by token (no auth)",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file returned"
          },
          "404": {
            "description": "Receipt not found or revoked"
          }
        },
        "tags": [
          "public-receipts"
        ]
      }
    },
...
  },
  "info": {
    "title": "ReceiptMaster API",
    "description": "Multi-tenant SaaS API for managing products, recipients, orders, receipts, and analytics.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Recipient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "telegram_user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ]
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "order_id": {
            "type": "string"
          },
          "product_id": {
            "type": "string"
          },
          "product_name": {
            "type": "string"
          },
          "unit_price_cents": {
            "type": "number"
          },
          "qty": {
            "type": "number"
          },
          "line_total_cents": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "order_id",
          "product_id",
          "product_name",
          "unit_price_cents",
          "qty",
          "line_total_cents"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "recipient_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "confirmed",
              "cancelled"
            ]
          },
          "subtotal_cents": {
            "type": "number"
          },
          "total_cents": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          },
          "is_locked": {
            "type": "boolean"
          },
          "recipient": {
            "$ref": "#/components/schemas/Recipient"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "receipts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Receipt"
            }
          }
        },
        "required": [
          "id",
          "recipient_id",
          "status",
          "subtotal_cents",
          "total_cents",
          "currency",
          "created_by",
          "created_at",
          "updated_at",
          "is_locked",
          "recipient",
          "items",
          "receipts"
        ]
      },
      "Receipt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "order_id": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "pdf_url": {
            "type": "string"
          },
          "pdf_path": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "generated",
              "void"
            ]
          },
          "html_snapshot": {
            "type": "string"
          },
          "template_id": {
            "type": "string"
          },
          "template_version": {
            "type": "number"
          },
          "voided_at": {
            "format": "date-time",
            "type": "string"
          },
          "void_reason": {
            "type": "string"
          },
          "public_token": {
            "type": "string"
          },
          "progress": {
            "type": "number"
          },
          "error_message": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        },
        "required": [
          "id",
          "order_id",
          "number",
          "status",
          "progress",
          "created_at",
          "order"
        ]
      },
...
      "PaginatedReceiptResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Receipt"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "data"
        ]
      },
...
    }
  }
}